[ft_irc] refactor: reorganization of files
