+++ src/commands/commandMode.cpp
@@ -6,7 +6,7 @@
 /*   By: macarval <macarval@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/23 13:19:18 by macarval          #+#    #+#             */
-/*   Updated: 2024/07/26 09:01:20 by macarval         ###   ########.fr       */
+/*   Updated: 2024/08/05 10:25:07 by macarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,18 +51,6 @@ void Commands::commandModeChannel(std::string &channelName)
 	}
 }
 
-bool Commands::verifyMode(std::string &mode, std::string &who,
-						  std::string pattern)
-{
-	if (mode[0] != '+' && mode[0] != '-')
-		printInfo(errorNeedMoreParams());
-	else if (mode.find_first_not_of(pattern) != std::string::npos)
-		printInfo(errorUnknownMode(who, mode));
-	else
-		return true;
-	return false;
-}
-
 void Commands::applyMode(std::string &channelName, std::string mode,
 						 size_t &index)
 {
@@ -87,39 +75,6 @@ void Commands::applyMode(std::string &channelName, std::string mode,
 	handleModeParameters(mode, channelName, index);
 }
 
-void Commands::handleModeParameters(std::string &mode,
-									std::string &channelName, size_t &index)
-{
-	Channel	&channel = _channels.get(channelName)->second;
-	std::string	param = _args[index];
-
-	if (mode.find_first_not_of("+-oOv") == std::string::npos)
-	{
-		std::map<int, Client>::iterator it = _clients.getClientByNick(param);
-		if (it == _clients.end())
-			printInfo(errorNoSuchNick(param));
-		else if (!it->second.isInChannel(channelName))
-			printInfo(errorUserNotInChannel(param, channelName));
-		else
-			channel.setUserMode(_clients.getFDByNick(param), mode);
-	}
-	else if (mode == "+k")
-		handleKeyMode(channelName, param);
-	else if (mode == "+l")
-		channel.setUserLimit(toInt(param));
-	++index;
-}
-
-void Commands::handleKeyMode(std::string &channelName, std::string &param)
-{
-	Channel	&channel = _channels.get(channelName)->second;
-
-	if (channel.getKey().empty())
-		channel.setKey(param);
-	else
-		printInfo(errorKeySet(channelName));
-}
-
 void Commands::commandModeUser(std::string &nick)
 {
 	if (validArg(nick) && verifyNick(nick))
@@ -140,18 +95,4 @@ void Commands::commandModeUser(std::string &nick)
 			}
 		}
 	}
-}
-
-bool Commands::verifyNick(std::string &nick)
-{
-	Client client = _clients.getClient(_fd)->second;
-
-	if (_clients.getClientByNick(nick) == _clients.end())
-		printInfo(errorNoSuchNick(nick));
-	else if (nick != client.getNick()
-			 && !client.getMode(Client::OPERATOR))
-		printInfo(errorUsersDontMatch());
-	else
-		return true;
-	return false;
 }
