+++ src/commands/utils/errorsCode.cpp
