+++ src/commands/utils/infos.cpp
