+++ src/commands/utils/validations.cpp
