+++ src/commands/utils/verify.cpp
